+++ src/index.js
@@ -1,29 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { applyMiddleware, createStore } from 'redux';
 
-import { promiseMiddleware } from './middleware';
+import store from './store';
 
 import App from './components/App';
-
-const initialState = {
-  appName: 'Medium Lite',
-  articles: null
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case 'HOME_PAGE_LOADED':
-      return {
-        ...state,
-        articles: action.payload.articles
-      };
-  }
-  return state;
-};
-
-const store = createStore(reducer, applyMiddleware(promiseMiddleware));
 
 ReactDOM.render((
   <Provider store={store}>
